Rename svg icons
