+++ client/src/components/AppCard.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
 
-import appIconPlaceholder from '../app-icon-placeholder.svg';
+import appIconPlaceholder from '../images/logoPlaceholder.svg';
 import './AppCard.scss';
 
 const AppCard = ({
